Lib: introduce a new library js_of_ocaml-compiler.runtime to host core primitives
