+++ lib/js_of_ocaml/dune
@@ -1,7 +1,7 @@
 (library
  (name js_of_ocaml)
  (public_name js_of_ocaml)
- (libraries uchar bytes)
+ (libraries uchar bytes js_of_ocaml-compiler.runtime)
  (foreign_stubs
   (language c)
   (names js_of_ocaml_stubs))
@@ -11,12 +11,13 @@
 (rule
  (targets js_of_ocaml_stubs.c)
  (deps
-  (glob_files *.ml))
+  (:jsoo (glob_files *.ml))
+  (:runtime (glob_files ../runtime/*.ml)))
  (mode promote)
  (action
   (with-stdout-to
    %{targets}
-   (run ../gen_stubs/gen_stubs.exe %{deps}))))
+   (run ../gen_stubs/gen_stubs.exe %{jsoo} --except %{runtime}))))
 
 (rule
  (targets lib_version.ml)
