+++ lib/js_of_ocaml/js_of_ocaml_stubs.c
@@ -1,91 +1,11 @@
 #include <stdlib.h>
 #include <stdio.h>
-void bigstring_of_array_buffer () {
-  fprintf(stderr, "Unimplemented Javascript primitive bigstring_of_array_buffer!\n");
-  exit(1);
-}
-void bigstring_of_typed_array () {
-  fprintf(stderr, "Unimplemented Javascript primitive bigstring_of_typed_array!\n");
-  exit(1);
-}
-void bigstring_to_array_buffer () {
-  fprintf(stderr, "Unimplemented Javascript primitive bigstring_to_array_buffer!\n");
-  exit(1);
-}
-void bigstring_to_typed_array () {
-  fprintf(stderr, "Unimplemented Javascript primitive bigstring_to_typed_array!\n");
-  exit(1);
-}
-void caml_ba_from_typed_array () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_ba_from_typed_array!\n");
-  exit(1);
-}
-void caml_ba_kind_of_typed_array () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_ba_kind_of_typed_array!\n");
-  exit(1);
-}
-void caml_ba_to_typed_array () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_ba_to_typed_array!\n");
-  exit(1);
-}
-void caml_create_file () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_create_file!\n");
-  exit(1);
-}
-void caml_exn_with_js_backtrace () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_exn_with_js_backtrace!\n");
-  exit(1);
-}
 void caml_int64_create_lo_mi_hi () {
   fprintf(stderr, "Unimplemented Javascript primitive caml_int64_create_lo_mi_hi!\n");
   exit(1);
 }
-void caml_js_call () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_js_call!\n");
-  exit(1);
-}
-void caml_js_delete () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_js_delete!\n");
-  exit(1);
-}
-void caml_js_equals () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_js_equals!\n");
-  exit(1);
-}
 void caml_js_error_of_exception () {
   fprintf(stderr, "Unimplemented Javascript primitive caml_js_error_of_exception!\n");
-  exit(1);
-}
-void caml_js_eval_string () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_js_eval_string!\n");
-  exit(1);
-}
-void caml_js_expr () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_js_expr!\n");
-  exit(1);
-}
-void caml_js_from_array () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_js_from_array!\n");
-  exit(1);
-}
-void caml_js_from_bool () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_js_from_bool!\n");
-  exit(1);
-}
-void caml_js_from_float () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_js_from_float!\n");
-  exit(1);
-}
-void caml_js_from_string () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_js_from_string!\n");
-  exit(1);
-}
-void caml_js_fun_call () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_js_fun_call!\n");
-  exit(1);
-}
-void caml_js_get () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_js_get!\n");
   exit(1);
 }
 void caml_js_get_console () {
@@ -100,155 +20,15 @@ void caml_js_html_escape () {
   fprintf(stderr, "Unimplemented Javascript primitive caml_js_html_escape!\n");
   exit(1);
 }
-void caml_js_instanceof () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_js_instanceof!\n");
-  exit(1);
-}
-void caml_js_meth_call () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_js_meth_call!\n");
-  exit(1);
-}
-void caml_js_new () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_js_new!\n");
-  exit(1);
-}
-void caml_js_object () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_js_object!\n");
-  exit(1);
-}
 void caml_js_on_ie () {
   fprintf(stderr, "Unimplemented Javascript primitive caml_js_on_ie!\n");
-  exit(1);
-}
-void caml_js_pure_expr () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_js_pure_expr!\n");
-  exit(1);
-}
-void caml_js_set () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_js_set!\n");
-  exit(1);
-}
-void caml_js_to_array () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_js_to_array!\n");
-  exit(1);
-}
-void caml_js_to_bool () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_js_to_bool!\n");
-  exit(1);
-}
-void caml_js_to_byte_string () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_js_to_byte_string!\n");
-  exit(1);
-}
-void caml_js_to_float () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_js_to_float!\n");
-  exit(1);
-}
-void caml_js_to_string () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_js_to_string!\n");
-  exit(1);
-}
-void caml_js_typeof () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_js_typeof!\n");
   exit(1);
 }
 void caml_js_var () {
   fprintf(stderr, "Unimplemented Javascript primitive caml_js_var!\n");
   exit(1);
 }
-void caml_js_wrap_callback () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_js_wrap_callback!\n");
-  exit(1);
-}
-void caml_js_wrap_callback_arguments () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_js_wrap_callback_arguments!\n");
-  exit(1);
-}
-void caml_js_wrap_callback_strict () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_js_wrap_callback_strict!\n");
-  exit(1);
-}
-void caml_js_wrap_meth_callback () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_js_wrap_meth_callback!\n");
-  exit(1);
-}
-void caml_js_wrap_meth_callback_arguments () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_js_wrap_meth_callback_arguments!\n");
-  exit(1);
-}
-void caml_js_wrap_meth_callback_strict () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_js_wrap_meth_callback_strict!\n");
-  exit(1);
-}
-void caml_js_wrap_meth_callback_unsafe () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_js_wrap_meth_callback_unsafe!\n");
-  exit(1);
-}
-void caml_jsbytes_of_string () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_jsbytes_of_string!\n");
-  exit(1);
-}
-void caml_jsstring_of_string () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_jsstring_of_string!\n");
-  exit(1);
-}
-void caml_list_mount_point () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_list_mount_point!\n");
-  exit(1);
-}
-void caml_list_of_js_array () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_list_of_js_array!\n");
-  exit(1);
-}
-void caml_list_to_js_array () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_list_to_js_array!\n");
-  exit(1);
-}
-void caml_ml_set_channel_output () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_ml_set_channel_output!\n");
-  exit(1);
-}
-void caml_ml_set_channel_refill () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_ml_set_channel_refill!\n");
-  exit(1);
-}
-void caml_mount_autoload () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_mount_autoload!\n");
-  exit(1);
-}
-void caml_ojs_new_arr () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_ojs_new_arr!\n");
-  exit(1);
-}
-void caml_pure_js_expr () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_pure_js_expr!\n");
-  exit(1);
-}
-void caml_read_file_content () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_read_file_content!\n");
-  exit(1);
-}
-void caml_string_of_array () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_string_of_array!\n");
-  exit(1);
-}
-void caml_string_of_jsbytes () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_string_of_jsbytes!\n");
-  exit(1);
-}
-void caml_string_of_jsstring () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_string_of_jsstring!\n");
-  exit(1);
-}
-void caml_unmount () {
-  fprintf(stderr, "Unimplemented Javascript primitive caml_unmount!\n");
-  exit(1);
-}
 void caml_xmlhttprequest_create () {
   fprintf(stderr, "Unimplemented Javascript primitive caml_xmlhttprequest_create!\n");
-  exit(1);
-}
-void debugger () {
-  fprintf(stderr, "Unimplemented Javascript primitive debugger!\n");
   exit(1);
 }
